refactor: make smaller forum and topics components and move them into features and widgets folders
