+++ packages/client/src/features/AddForumTopic/index.ts
@@ -1,0 +1,1 @@
+export { AddTopicForm } from './ui/AddTopicForm';
