+++ packages/client/src/features/AddForumTopicComment/index.ts
@@ -1,0 +1,1 @@
+export { AddTopicCommentForm } from './ui/AddTopicCommentForm';
