+++ packages/client/src/shared/constants/forum/types/index.ts
@@ -1,0 +1,4 @@
+export type { TAuthor, TTopicComment, TTopicInfo } from './Topic.models';
+export { TopicInfo } from './Topic.models';
+export type { TForumItem, TAddTopicFormValues } from './Forum.models';
+export { ADD_FORUM_FORM_ID } from './Forum.models';
