+++ packages/client/src/widgets/ForumTopicComments/index.ts
@@ -1,0 +1,1 @@
+export { ForumTopicComments } from './ui/ForumTopicComments';
