+++ packages/client/src/widgets/ForumTopicComments/ui/ForumTopicComments.css
@@ -1,0 +1,23 @@
+.topic__comments {
+  margin-top: 24px;
+  padding-left: 8px;
+}
+
+.topic__comments-title.ant-typography {
+  font-size: 18px;
+  font-weight: 600;
+}
+
+.topic__comments-content {
+  width: 100%;
+  font-size: 16px;
+  max-height: 100px;
+  overflow-y: auto;
+  margin-top: 12px;
+}
+
+.topic__info-text.ant-typography {
+  margin-bottom: 0;
+  line-height: 14px;
+  font-size: 16px;
+}
