+++ packages/client/src/widgets/ForumTopicHeader/index.ts
@@ -1,0 +1,1 @@
+export { ForumTopicHeader } from './ui/ForumTopicHeader';
