+++ packages/client/src/widgets/ForumTopicsTable/index.ts
@@ -1,0 +1,1 @@
+export { ForumTopicsTable } from './ui/ForumTopicsTable';
